python/requirements.txt
dspy-ai>=2.4.0
openai>=1.0.0,<2.0.0
